Refactor component structure and update imports
- Moved GenerateTheoryModal to dialogs directory and updated import paths accordingly.
- Refactored MultiSelect and Select components into a dedicated select directory.
- Updated PageContainer and Sidebar components to a layout directory.
- Adjusted imports in various components and routes to reflect new structure.
- Removed unused components: GenerateTheoryModal, LanguageSelect, MultiSelect, Select, PageContainer, Sidebar, and Topbar.
- Enhanced the GenerateTheoryModal with improved file selection and UI elements.
