+++ frontend/src/lib/components/layout/Sidebar.svelte
