+++ frontend/src/routes/(app)/learningkit/[learningKitId]/edit-form/+page.svelte
@@ -2,7 +2,7 @@
 	import SuperDebug, { superForm } from 'sveltekit-superforms';
 	import { toaster } from '$lib/states/toasterState.svelte.js';
 	import { useQueryInvalidation } from '$lib/api/useQueryInvalidation.js';
-	import PageContainer from '$lib/components/PageContainer.svelte';
+	import PageContainer from '$lib/components/layout/PageContainer.svelte';
 	import { _ } from 'svelte-i18n';
 	import { page } from '$app/state';
